Finialized the pet feature
