+++ ProefProeveVeluwe/Assets/Scripts/Runtime/Animal Interaction.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7682b23894caa11488b3ae4b7ada65f8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
