+++ ProefProeveVeluwe/Assets/Scripts/Runtime/Animal Interaction/Petting.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d589e5450aab5d44fbf354c92dc1e89a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
