+++ ProefProeveVeluwe/ProjectSettings/EditorBuildSettings.asset
@@ -6,9 +6,6 @@ EditorBuildSettings:
   serializedVersion: 2
   m_Scenes:
   - enabled: 1
-    path: Assets/Scenes/Minigames/PetDeer.unity
-    guid: dc531286bf607ca4f96eeda1826ac6a1
-  - enabled: 1
     path: Assets/Scenes/Test/GPSDebugTest.unity
     guid: 99c9720ab356a0642a771bea13969a05
   m_configObjects: {}
